fixed reporoot, homedir and -m UI and -n 2

+++ homework-2/conftest.py
@@ -1,6 +1,7 @@
 from ui.fixtures import *
 
 import os
+import sys
 import shutil
 import pytest
 
@@ -16,21 +17,24 @@ def config(request):
     return {'browser': browser, 'url': url}
 
 
-@pytest.fixture(scope='session')
-def base_temp_dir():
+# @pytest.fixture(scope='session')
+def pytest_configure(config):
 
-    base_dir = '/home/dashori/Рабочий стол/maiil/tests'
+    if sys.platform.startswith('win'):
+        base_dir = 'C:\\tests'
+    else:
+        base_dir = '/tmp/tests'
+   
+    if not hasattr(config, 'workerinput'):
+        if os.path.exists(base_dir):
+            shutil.rmtree(base_dir)
 
-    if os.path.exists(base_dir):
-        shutil.rmtree(base_dir)
+        os.makedirs(base_dir)
 
-    os.makedirs(base_dir)
-    return base_dir
-
+    config.base_temp_dir = base_dir
 
 @pytest.fixture(scope='function')
-def temp_dir(base_temp_dir, request):
-    test_dir = os.path.join(base_temp_dir, request._pyfuncitem.nodeid.replace('/', '_').replace(':', '_'))
+def temp_dir(request):
+    test_dir = os.path.join(request.config.base_temp_dir, request._pyfuncitem.nodeid.replace('/', '_').replace(':', '_'))
     os.makedirs(test_dir)
     return test_dir
-
